test(vetting): Wave 10 — mechanics/fixture renames (§6.3)
Renamed the I/O + plumbing tests to the _mechanics suffix (git mv,
headers only; test function names and TEST() registrations unchanged):
test_initialization, test_roi_blacklist, test_tiff_loader, test_3d_nifti,
test_omezarr, test_arrow, test_arrow_file_name -> *_mechanics.h. Only the
seven #include lines in test_all.cc changed. Pure fixtures/harness left
unrenamed per §6.3. None appear in the registry current_test, so no
repoint needed. (Arrow files are under #ifdef USE_ARROW = OFF locally, so
correct-by-inspection but not compile-exercised.)

All C++ test files now follow the taxonomy.

VERIFIED: full gtest suite 696/696 — no tests dropped; registry rc=0.
