+++ tests/test_roi_blacklist_mechanics.h
