+++ tests/test_tiff_loader_mechanics.h
